feat(kuma-cp) resolve review comments
