+++ .circleci/config.yml
@@ -351,7 +351,7 @@ jobs:
         # otherwise, fallback to the most recently generated cache on any branch
         - go.mod-
     - run:
-        name: Build all Kuma binaries (such as, kumactl, kuma-cp, kuma-dp, kuma-injector, kuma-prometheus-sd)
+        name: Build all Kuma binaries (such as, kumactl, kuma-cp, kuma-dp, kuma-prometheus-sd)
         command: make build
 
     # Persist the specified paths into the workspace for use in downstream jobs
