+++ app/kumactl/data/install/k8s/control-plane/kuma-cp/app.yaml
@@ -63,40 +63,6 @@ data:
     #
     # see conf/kuma-cp.conf for available settings
 ---
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: kuma-injector-config
-  namespace: {{ .Namespace }}
-data:
-  config.yaml: |
-    injector:
-      sidecarContainer:
-        redirectPort: 15001
-        uid: 5678
-        gid: 5678
-        adminPort: 9901
-        drainTime: 30s
-
-        readinessProbe:
-          initialDelaySeconds: 1
-          timeoutSeconds:      3
-          periodSeconds:       5
-          successThreshold:    1
-          failureThreshold:    12
-        livenessProbe:
-          initialDelaySeconds: 60
-          timeoutSeconds:      3
-          periodSeconds:       5
-          failureThreshold:    12
-        resources:
-          requests:
-            cpu: 50m
-            memory: 64Mi
-          limits:
-            cpu: 1000m
-            memory: 512Mi
----
 apiVersion: apps/v1
 kind: Deployment
 metadata:
@@ -223,32 +189,26 @@ webhooks:
     - UPDATE
     resources:
     - meshes
----
-apiVersion: admissionregistration.k8s.io/v1beta1
-kind: MutatingWebhookConfiguration
-metadata:
-  name: kuma-injector-webhook-configuration
-webhooks:
-  - name: kuma-injector.kuma.io
-    namespaceSelector:
-      matchLabels:
-        kuma.io/sidecar-injection: enabled
-    failurePolicy: {{ .InjectorFailurePolicy }}
-    clientConfig:
-      caBundle: {{ .AdmissionServerTlsCert | b64enc }}
-      service:
-        namespace: {{ .Namespace }}
-        name: {{ .ControlPlaneServiceName }}
-        path: /inject-sidecar
-    rules:
-      - apiGroups:
-          - ""
-        apiVersions:
-          - v1
-        operations:
-          - CREATE
-        resources:
-          - pods
+- name: kuma-injector.kuma.io
+  namespaceSelector:
+    matchLabels:
+      kuma.io/sidecar-injection: enabled
+  failurePolicy: {{ .InjectorFailurePolicy }}
+  clientConfig:
+    caBundle: {{ .AdmissionServerTlsCert | b64enc }}
+    service:
+      namespace: {{ .Namespace }}
+      name: {{ .ControlPlaneServiceName }}
+      path: /inject-sidecar
+  rules:
+    - apiGroups:
+        - ""
+      apiVersions:
+        - v1
+      operations:
+        - CREATE
+      resources:
+        - pods
 ---
 apiVersion: admissionregistration.k8s.io/v1beta1
 kind: ValidatingWebhookConfiguration
