+++ pkg/config/plugins/runtime/k8s/testdata/default-config.golden.yaml
@@ -4,9 +4,6 @@ admissionServer:
   port: 5443
 injector:
   cniEnabled: false
-  controlPlane:
-    apiServer:
-      url: http://kuma-control-plane.kuma-system:5681
   sidecarContainer:
     image: kuma/kuma-dp:latest
     redirectPort: 15001
