+++ pkg/plugins/runtime/k8s/plugin.go
@@ -1,14 +1,8 @@
 package k8s
 
 import (
-	"context"
-	"encoding/json"
 	"fmt"
-	"net/http"
 	"strings"
-
-	kube_core "k8s.io/api/core/v1"
-	kube_admission "sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 
 	"github.com/Kong/kuma/pkg/plugins/runtime/k8s/webhooks/injector"
 
@@ -148,58 +142,10 @@ func addValidators(mgr kube_ctrl.Manager, rt core_runtime.Runtime) error {
 }
 
 func addMutators(mgr kube_ctrl.Manager, rt core_runtime.Runtime) {
-	mgr.GetWebhookServer().Register("/inject-sidecar",
-		PodMutatingWebhook(injector.New(rt.Config().Runtime.Kubernetes.Injector, mgr.GetClient()).InjectKuma))
-}
-
-//func Setup(mgr kube_manager.Manager, cfg *kuma_injector_conf.Config) error {
-//	webhookServer := &kube_webhook.Server{
-//		Host:    cfg.WebHookServer.Address,
-//		Port:    int(cfg.WebHookServer.Port),
-//		CertDir: cfg.WebHookServer.CertDir,
-//	}
-//	if err := mesh_k8s.AddToScheme(mgr.GetScheme()); err != nil {
-//		return errors.Wrap(err, "could not add to scheme")
-//	}
-//
-//	if err := k8scnicncfio.AddToScheme(mgr.GetScheme()); err != nil {
-//		return errors.Wrap(err, "could not add to scheme")
-//	}
-//
-//	webhookServer.Register("/inject-sidecar", PodMutatingWebhook(injector.New(cfg.Injector, mgr.GetClient()).InjectKuma))
-//	webhookServer.WebhookMux.HandleFunc("/healthy", func(resp http.ResponseWriter, _ *http.Request) {
-//		resp.WriteHeader(http.StatusOK)
-//	})
-//	webhookServer.WebhookMux.HandleFunc("/ready", func(resp http.ResponseWriter, _ *http.Request) {
-//		resp.WriteHeader(http.StatusOK)
-//	})
-//	return mgr.Add(webhookServer)
-//}
-
-type PodMutator func(*kube_core.Pod) error
-
-func PodMutatingWebhook(mutator PodMutator) *kube_admission.Webhook {
-	return &kube_admission.Webhook{
-		Handler: &podMutatingHandler{mutator: mutator},
-	}
-}
-
-type podMutatingHandler struct {
-	mutator PodMutator
-}
-
-func (h *podMutatingHandler) Handle(ctx context.Context, req kube_webhook.AdmissionRequest) kube_webhook.AdmissionResponse {
-	//webhookLog.V(1).Info("received request", "request", req)
-	var pod kube_core.Pod
-	if err := json.Unmarshal(req.Object.Raw, &pod); err != nil {
-		return kube_admission.Errored(http.StatusBadRequest, err)
-	}
-	if err := h.mutator(&pod); err != nil {
-		return kube_admission.Errored(http.StatusInternalServerError, err)
-	}
-	mutatedRaw, err := json.Marshal(pod)
-	if err != nil {
-		return kube_admission.Errored(http.StatusInternalServerError, err)
-	}
-	return kube_admission.PatchResponseFromRaw(req.Object.Raw, mutatedRaw)
+	kumaInjector := injector.New(
+		rt.Config().Runtime.Kubernetes.Injector,
+		rt.Config().ApiServer.Catalog.ApiServer.Url,
+		mgr.GetClient(),
+	)
+	mgr.GetWebhookServer().Register("/inject-sidecar", k8s_webhooks.PodMutatingWebhook(kumaInjector.InjectKuma))
 }
