--- tools/releases/dockerfiles/Dockerfile.kuma-injector
@@ -1,12 +1,0 @@
-FROM alpine:3.10
-
-ADD $KUMA_ROOT/build/artifacts-linux-amd64/kuma-injector/kuma-injector /usr/bin
-
-RUN mkdir /kuma
-COPY $KUMA_ROOT/tools/releases/templates/LICENSE /kuma
-COPY $KUMA_ROOT/tools/releases/templates/NOTICE /kuma
-COPY $KUMA_ROOT/tools/releases/templates/README /kuma
-
-USER nobody:nobody
-
-ENTRYPOINT ["kuma-injector"]
